src/main/java/com/example/demo/GameController.java
package com.example.demo;

import org.springframework.stereotype.Controller;
import org.springframework.ui.Model;
import org.springframework.ui.ModelMap;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestParam;

@Controller
public class GameController {
    private final GameService gameService;

    public GameController(GameService gameService) {
        this.gameService = gameService;
    }

    @GetMapping("/game")
    public String game(@RequestParam(required = false) Integer guess, ModelMap model) {
        if (guess != null) {
            model.addAttribute("message", gameService.checkGuess(guess));
        }
        return "game";
    }
}